// src/types/Character.ts
export interface Character {
    "url": string;
    "name": string;
    "gender": string;
    "culture": string;
    "born": string;
    "died": string;
    "titles": Array<string>;
    "aliases": Array<string>;
    "father": string;
    "mother": string;
    "spouse": string;
    "allegiances": Array<string>;
    "books": Array<string>;
    "povBooks": Array<string>;
    "tvSeries": Array<string>;
    "playedBy": Array<string>;
}